fin multi ray sphere
