+++ files/basic.rt
@@ -3,4 +3,6 @@ A   0.2	                                    255,255,255
 C   0,0,0        0,0,1          70
 L   5,5,10                      1.0         255,255,255
 
-sp  -1.5,0,0                      2            0,255,0
+sp  -1.5,0,0                      0.5             255,0,0
+sp  0,0,0                         1               0,255,0
+sp  1.5,0,0                      0.75             0,0,255
