+++ includes/utils.h
@@ -23,11 +23,12 @@ t_vtr	vtrcross(t_vtr a, t_vtr b);
 void	vtrprint(t_vtr v);
 // * matrix
 float	**mtx_identity(int size);
-float	**mtx_multi(float **m1, float **m2, int size);
-void	mtx_print(float **m, int size);
-float	mtx_det(float **m, int size);
+float	**mtx_trans(float **m, int size);
 float	**mtx_sub(float **m, int size, int row, int col);
-void	mtx_free(float **m, int size);
+float	**mtx_multi(float **m1, float **m2, int size);
 float	**mtx_inverse(float **m, int size);
+float	mtx_det(float **m, int size);
+void	mtx_print(float **m, int size);
+void	mtx_free(float **m, int size);
 
 #endif
