+++ srcs/utils/matrix1.c
@@ -25,6 +25,24 @@ float	**mtx_identity(int size)
 	return (m);
 }
 
+float	**mtx_trans(float **m, int size)
+{
+	float	**mm;
+	int		i;
+	int		j;
+
+	mm = ft_calloc(size, sizeof(float *));
+	i = -1;
+	while (++i < size)
+	{
+		j = -1;
+		mm[i] = ft_calloc(size, sizeof(float));
+		while (++j < size)
+			mm[i][j] = m[j][i];
+	}
+	return (mm);
+}
+
 float	**mtx_sub(float **m, int size, int row, int col)
 {
 	float	**mm;
@@ -86,6 +104,7 @@ float	**mtx_inverse(float **m, int size)
 	float	**mm;
 	float	**subm;
 	float	det;
+	float	**mmt;
 	int	i;
 	int	j;
 
@@ -103,58 +122,7 @@ float	**mtx_inverse(float **m, int size)
 			mtx_free(subm, size - 1);
 		}
 	}
-	return (mm);
-}
-
-float	mtx_det(float **m, int size)
-{
-	int i;
-	int	sign;
-	float sum;
-	if (size == 2)
-		return (m[0][0] * m[1][1] - m[0][1] * m[1][0]);
-	i = -1;
-	sign = 1;
-	sum = 0.0;
-	while (++i < size)
-	{
-		float **subm = mtx_sub(m, size, 0, i);
-		sum += m[0][i] * sign * mtx_det(subm, size - 1);
-		mtx_free(subm, size - 1);
-		sign = -sign;
-	}
-	return (sum);
-}
-
-
-void	mtx_print(float **m, int size)
-{
-	int	i;
-	int	j;
-
-	i = -1;
-	printf("{\n");
-	while (++i < size)
-	{
-		j = -1;
-		printf("  { ");
-		while (++j < size)
-		{
-			printf("%f", m[i][j]);
-			if (j != size - 1)
-				printf(", ");
-		}
-		printf(" }\n");
-	}
-	printf("}\n");
-}
-
-void	mtx_free(float **m, int size)
-{
-	int	i;
-
-	i = -1;
-	while (++i < size)
-		free(m[i]);
-	free(m);
+	mmt = mtx_trans(mm, 4);
+	mtx_free(mm, 4);
+	return (mmt);
 }
